mysterioese compilefehler auf anderen System hoffentlich beseitigt

+++ src/core/charconv.cpp
@@ -1,6 +1,7 @@
 #include "charconv.h"
 #include "debug.h"
 #include "world.h"
+#include <algorithm>
 
 CharConv::CharConv()
 	: m_bitstream()
@@ -81,7 +82,7 @@ void CharConv::toBuffer(std::string s, unsigned int size)
 	{
 		char * data = new char[size];
 		memset(data,0,size);
-		memcpy(data,s.data(), std::min(size,s.size()));
+		memcpy(data,s.data(), std::min(size,(unsigned int) s.size()));
 		toBuffer(data, size);
 		delete data;
 	}
